refine (and rename) the bertin153 facet projection test
